requirements.txt
# Core AI/LLM Libraries
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Web Framework & HTTP Requests
streamlit>=1.28.0
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Data Handling & Validation
pydantic>=2.0.0
python-dateutil>=2.8.0

# Optional but useful utilities
typing-extensions>=4.7.0 